% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litr.R
\name{do_not_edit_message}
\alias{do_not_edit_message}
\title{Generate do-not-edit message to put at top of file}
\usage{
do_not_edit_message(rmd_file, type = c("R", "man"))
}
\arguments{
\item{rmd_file}{Name of the Rmd file to mention}

\item{type}{Whether this is a R/ file or a man/ file}
}
\description{
Generate do-not-edit message to put at top of file
}
