% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litr.R
\name{document}
\alias{document}
\title{Use roxygen to document a package}
\usage{
document(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{devtools::document()}}
}
\description{
This is a wrapper for the \code{devtools::document()} function, which in turn is a
wrapper for the \code{roxygen2::roxygenize()} function.  The purpose for \code{litr}
having this wrapper is to make one modification.  In particular, the line
in the outputted \code{Rd} files should not say "Please edit documentation in
R/file.R" but instead should refer to the Rmd file that generates everything.
}
