% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liter.R
\name{send_to_package}
\alias{send_to_package}
\title{A knitr Chunk Hook for Writing R Code and Tests}
\usage{
send_to_package(before, options, envir)
}
\arguments{
\item{before}{Indicates whether this is being called before or after the
chunk code is executed}

\item{options}{Has information from the chunk}

\item{envir}{Environment}
}
\description{
This chunk hook detects whether a chunk is defining a function to be included
in the R package (looks for the Roxygen2 comment format \verb{#' }).  If so, then it
is written to the \verb{R/} directory.  It also looks for chunks with \verb{testthat::}
in them, which are written to the \code{tests} directory of the R package.
}
