% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litr.R
\name{setup}
\alias{setup}
\title{Code for setup chunk}
\usage{
setup(package_name)
}
\arguments{
\item{package_name}{Name of R package to create}
}
\description{
Creates directory where package will be. (Deletes what is currently there as
long as it appears to have been created by the Rmd file from which this
function is being called.)  Sets the root directory to this directory and
sets up the main chunk hook \code{litr::send_to_package} that sends code to the R
package directory.
}
