% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Code for setup chunk}
\usage{
setup(package_dir)
}
\arguments{
\item{package_dir}{Directory where R package will be created}
}
\description{
Creates directory where package will be. (Deletes what is currently there as
long as it appears to have been created by litr and does not have any
subsequent manual edits.)  Sets the root directory to this directory and
sets up the main chunk hook \code{litr::send_to_package} that sends code to the R
package directory.
}
